Multi-Modal-Gradio-Assistant/requirements.txt
ollama
openai
openai-whisper
python-dotenv
jupyterlab
ipywidgets
streamlit
anthropic
google.generativeai
gradio